creates config file

+++ src/tasks/tasks.service.ts
@@ -1,35 +1,34 @@
-import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTaskDto } from './dto/get-task.dto';
-import { stat } from 'fs';
 import { TaskRepository } from './task.repository';
-import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
   constructor(private readonly taskRepository: TaskRepository) {}
 
-  getTasks(getTaskDto: GetTaskDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(getTaskDto);
+  getTasks(getTaskDto: GetTaskDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(getTaskDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne({ where: { id } });
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const task = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
-    if (!found) {
+    if (!task) {
       throw new NotFoundException(`Task with id ${id} not found`);
     }
 
-    return found;
+    return task;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return await this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return await this.taskRepository.createTask(createTaskDto, user);
   }
-  async deleteTask(id: number): Promise<string> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<string> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ${id} not found`);
@@ -37,8 +36,8 @@ export class TasksService {
 
     return 'Task deleted Succesfully';
   }
-  async updateTask(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTask(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
 
     if (!task) {
       throw new NotFoundException(`Task with ${id} not found`);
